+++ call.go
@@ -145,6 +145,13 @@ func (c *Call) Header(k string) string {
 	return (*encoding.Call)(c).Header(k)
 }
 
+// OriginalHeader returns the value of the given request header provided with the
+// request. The getter is suitable for transport like TChannel that hides
+// certain headers by default eg: the ones starting with $
+func (c *Call) OriginalHeader(k string) string {
+	return (*encoding.Call)(c).OriginalHeader(k)
+}
+
 // OriginalHeaders returns a copy of the given request headers provided with the request.
 // The header key are not canonicalized and suitable for case-sensitive transport like TChannel.
 func (c *Call) OriginalHeaders() map[string]string {