+++ yarpcerrors/yarpcerrorclassifier_test.go
@@ -21,8 +21,9 @@
 package yarpcerrors
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestGetFaultTypeFromErrorForClientErrors(t *testing.T) {